+++ src/project/environment.rs
@@ -5,13 +5,18 @@ use super::{
 };
 use crate::project::has_features::HasFeatures;
 
+use crate::consts;
 use crate::task::TaskName;
 use crate::{task::Task, Project};
 use itertools::Either;
-use rattler_conda_types::Platform;
-use std::collections::HashSet;
-use std::hash::{Hash, Hasher};
-use std::{collections::HashMap, fmt::Debug};
+use rattler_conda_types::{Arch, Platform};
+use std::{
+    collections::{HashMap, HashSet},
+    fmt::Debug,
+    fs,
+    hash::{Hash, Hasher},
+    sync::Once,
+};
 
 /// Describes a single environment from a project manifest. This is used to describe environments
 /// that can be installed and activated.
@@ -95,6 +100,47 @@ impl<'p> Environment<'p> {
         self.project
             .environments_dir()
             .join(self.environment.name.as_str())
+    }
+
+    /// Returns the best platform for the current platform & environment.
+    pub fn best_platform(&self) -> Platform {
+        let current = Platform::current();
+
+        // If the current platform is supported, return it.
+        if self.platforms().contains(&current) {
+            return current;
+        }
+
+        static WARN_ONCE: Once = Once::new();
+
+        // If the current platform is osx-arm64 and the environment supports osx-64, return osx-64.
+        if current.is_osx() && self.platforms().contains(&Platform::Osx64) {
+            WARN_ONCE.call_once(|| {
+                let warn_folder = self.project.pixi_dir().join(consts::ONE_TIME_MESSAGES_DIR);
+                let emulation_warn = warn_folder.join("macos-emulation-warn");
+                if !emulation_warn.exists() {
+                    tracing::warn!(
+                        "osx-arm64 (Apple Silicon) is not supported by the pixi.toml, falling back to osx-64 (emulated with Rosetta)"
+                    );
+                    // Create a file to prevent the warning from showing up multiple times. Also ignore the result.
+                    fs::create_dir_all(warn_folder).and_then(|_| {
+                        std::fs::File::create(emulation_warn)
+                    }).ok();
+                }
+            });
+            return Platform::Osx64;
+        }
+
+        if self.platforms().len() == 1 {
+            // Take the first platform and see if it is a WASM one.
+            if let Some(platform) = self.platforms().iter().next() {
+                if platform.arch() == Some(Arch::Wasm32) {
+                    return *platform;
+                }
+            }
+        }
+
+        current
     }
 
     /// Returns the tasks defined for this environment.
@@ -275,8 +321,6 @@ mod tests {
         );
     }
 
-    // TODO: Add a test to verify that feature specific channels work as expected.
-
     #[test]
     fn test_default_platforms() {
         let manifest = Project::from_str(
@@ -441,6 +485,104 @@ mod tests {
         assert_eq!(
             foo_env.activation_scripts(Some(Platform::Linux64)),
             vec!["foo.bat".to_string(), "linux.bat".to_string()]
+        );
+    }
+
+    #[test]
+    fn test_channel_feature_priority() {
+        let manifest = Project::from_str(
+            Path::new("pixi.toml"),
+            r#"
+        [project]
+        name = "foobar"
+        channels = ["a", "b"]
+        platforms = ["linux-64", "osx-64"]
+
+        [feature.foo]
+        channels = ["c", "d"]
+
+        [feature.bar]
+        channels = ["e", "f"]
+
+        [feature.barfoo]
+        channels = ["a", "f"]
+
+        [environments]
+        foo = ["foo"]
+        foobar = ["foo", "bar"]
+        barfoo = {features = ["barfoo"], no-default-feature=true}
+        "#,
+        )
+        .unwrap();
+
+        // All channels are added in order of the features and default is last
+        let foobar_channels = manifest.environment("foobar").unwrap().channels();
+        assert_eq!(
+            foobar_channels
+                .into_iter()
+                .map(|c| c.name.clone().unwrap())
+                .collect_vec(),
+            vec!["c", "d", "e", "f", "a", "b"]
+        );
+
+        let foo_channels = manifest.environment("foo").unwrap().channels();
+        assert_eq!(
+            foo_channels
+                .into_iter()
+                .map(|c| c.name.clone().unwrap())
+                .collect_vec(),
+            vec!["c", "d", "a", "b"]
+        );
+
+        // The default feature is not included in the channels, so only the feature channels are included.
+        let barfoo_channels = manifest.environment("barfoo").unwrap().channels();
+        assert_eq!(
+            barfoo_channels
+                .into_iter()
+                .map(|c| c.name.clone().unwrap())
+                .collect_vec(),
+            vec!["a", "f"]
+        )
+    }
+
+    #[test]
+    fn test_channel_feature_priority_with_redefinition() {
+        let manifest = Project::from_str(
+            Path::new("pixi.toml"),
+            r#"
+        [project]
+        name = "test"
+        channels = ["d", "a", "b"]
+        platforms = ["linux-64"]
+
+        [environments]
+        foo = ["foo"]
+
+        [feature.foo]
+        channels = ["a", "c", "b"]
+
+        "#,
+        )
+        .unwrap();
+
+        let foobar_channels = manifest.environment("default").unwrap().channels();
+        assert_eq!(
+            foobar_channels
+                .into_iter()
+                .map(|c| c.name.clone().unwrap())
+                .collect_vec(),
+            vec!["d", "a", "b"]
+        );
+
+        // Check if the feature channels are sorted correctly,
+        // and that the remaining channels from the default feature are appended.
+        let foo_channels = manifest.environment("foo").unwrap().channels();
+        assert_eq!(
+            foo_channels
+                .into_iter()
+                .map(|c| c.name.clone().unwrap())
+                .collect_vec(),
+            vec!["a", "c", "b", "d"]
         );
     }